syncing works now

+++ CollaborationBoard/Source/Point.cs
@@ -7,10 +7,10 @@ namespace CollaborationBoard
 {
     public class Point
     {
-        public int X { get; set; }
-        public int Y { get; set; }
+        public float X { get; set; }
+        public float Y { get; set; }
 
-        public Point(int x, int y)
+        public Point(float x, float y)
         {
             this.X = x;
             this.Y = y;
